Add files via upload
I have modified .htaccess file and create .htpasswrd for administrator folder password. so before we entered that folder it will show a pop up login like username and password first to prevent hacking.

the access information's are write inside the .htaccess file

so even the hackers input is right it will show errors, but its not.

After that the hackers it self have to know the index filename for that folder. You can named it like index.php or anything else that hard to guess ok.

So i use a unique and hard to guess for that index filename.
And i use anti sql injection too for otentication to database.

So it will be double or triple security maybe hehehe.

You are free to use it for security on your website.

Thanks.